Unify video export pipeline
